test: extract contract delegation into separate file
